+++ SpicesUpdate@claudiux/files/SpicesUpdate@claudiux/4.2/constants.js
@@ -18,7 +18,8 @@ function DEBUG() {
   return _debug.query_exists(null);
 };
 
-const DOWNLOAD_TIME = 2;
+const DOWNLOAD_TIME = 10;
+const SORT = "date";
 
 const APPLET_DIR = HOME_DIR + "/.local/share/cinnamon/applets/" + UUID;
 const SCRIPTS_DIR = APPLET_DIR + "/scripts";
@@ -157,6 +158,7 @@ module.exports = {
   DIR_MAP,
   DCONFCACHEUPDATED,
   DOWNLOAD_TIME,
+  SORT,
   _,
   EXP1, EXP2, EXP3,
   DEBUG,